+++ src/middleware/mod.rs
@@ -9,6 +9,7 @@ mod logger;
 
 pub mod cors;
 pub mod csrf;
+pub mod identity;
 mod defaultheaders;
 mod errhandlers;
 #[cfg(feature = "session")]